requirements.txt
argcomplete==1.12.3
asgiref==3.2.7
base58==0.2.3
boto3==1.17.109
botocore==1.20.109
certifi==2021.5.30
cffi==1.14.5
cfn-flip==1.2.3
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.6.0
django-summernote==0.8.11.6
djangorestframework==3.12.2
docutils==0.12
durationpy==0.5
future==0.18.2
futures==3.0.5
gunicorn==20.0.4
hjson==1.5.8
idna==2.10
importlib-metadata==4.6.1
jmespath==0.9.0
lambda-packages==0.5.0
mysqlclient==2.0.3
pep517==0.10.0
Pillow==7.1.1
pip-tools==6.2.0
placebo==0.9.0
protobuf==3.17.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.5.3
python-slugify==1.2.0
pytz==2019.3
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
six==1.10.0
sqlparse==0.3.1
toml==0.10.2
tqdm==4.7.6
troposphere==3.0.1
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
Werkzeug==0.11.10
whitenoise==5.0.1
wsgi-request-logger==0.4.5
zappa==0.53.0
zipp==3.5.0
